fill so that each line is at least n characters wide
After puzzling over this for a few days while sick, I found a solution for fill.
First linesDoc converts the Document from a tree of Concats to a list.
Next it constructs folds that list into a list of Documents each representing
one line. Finally fill joins the list of lines back into a whole doc filling
the head of each line with spaces if needed bring it up to the requested width.
This feels neater to me that the other solutions I was pursuing.